ctmap: Factor out metrics handling code into file
